[IMP] InteractiveTests: Moved sort interactive tests to ui helper tests
Few interactive sort test cases were written under plugins,
those are moved under helpers/ui since Interactive is helping with UI.

Task-2965421
